requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py36_0
cffi=1.14.0=py36h2e261b9_0
cudatoolkit=10.0.130=0
freetype=2.9.1=h8a8886c_1
intel-openmp=2020.1=217
jpeg=9b=h024ee3a_2
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py_0
openssl=1.1.1g=h7b6447c_0
pillow=7.1.2=py36hb39fc2d_0
pip=19.3.1=py36_0
pycparser=2.20=py_0
python=3.6.10=h0371630_0
pytorch=1.4.0=py3.6_cuda10.0.130_cudnn7.6.3_0
readline=7.0=h7b6447c_5
setuptools=44.0.0=py36_0
six=1.14.0=py36_0
sqlite=3.30.1=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py36_cu100
wheel=0.33.6=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
